Fixing looping import and bug on mobile
