+++ sudokuru/app/Components/SudokuBoard/Core/Functions/CellFunctions.ts
@@ -4,7 +4,10 @@ import {
   CellLocation,
 } from "../../../../Functions/LocalDatabase";
 import { HintObjectProps } from "../../SudokuBoard";
-import { isBoardDisabled } from "../../SudokuBoardFunctions";
+import {
+  doesCellHaveConflict,
+  isBoardDisabled,
+} from "../../SudokuBoardFunctions";
 import { isValueCorrect } from "./BoardFunctions";
 
 /**
@@ -57,26 +60,6 @@ export const areCellUpdatesDisabled = (
   } else {
     return false;
   }
-};
-
-/**
- * Checks if a given cell in the puzzle has a conflict with the solution.
- *
- * @param r - The row index of the cell.
- * @param c - The column index of the cell.
- * @param cell - The cell object containing its type and entry.
- * @returns True if the cell's entry is incorrect; false otherwise.
- */
-export const doesCellHaveConflict = (
-  sudokuBoard: SudokuObjectProps,
-  r: number,
-  c: number
-): boolean => {
-  const cell = sudokuBoard.puzzle[r][c];
-  if (cell.type == "note" || cell.entry == 0) {
-    return false;
-  }
-  return !(sudokuBoard.puzzle[r][c].entry === sudokuBoard.puzzleSolution[r][c]);
 };
 
 export const getSelectedCells = (
