+++ sudokuru/app/Components/SudokuBoard/SudokuBoardSharedFunctions.ts
@@ -1,0 +1,14 @@
+// This file is for shared functions outside of SudokuBoard component
+
+/**
+ * Returns the string converted to a title format i.e. replaces _ with spaces and capitalizes only the first letter of each word
+ * @param str - the string to convert
+ * @returns the converted string
+ */
+export const toTitle = (str: string) => {
+  return str
+    .toLowerCase()
+    .split("_")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(" ");
+};
